version 0.1, stable work clients, vehicles, contracts, invoices, catalog-items, calibration-protocols, audit-log

+++ src/main/java/com/example/protaxo/catalog/dto/CatalogItemResponse.java
@@ -8,6 +8,6 @@ public record CatalogItemResponse(
         CatalogItemType type,
         String name,
         BigDecimal basePrice,
-        BigDecimal stockQuantity
+        Integer stockQuantity
 ) {
 }
